drivers.c: Fix new warning.
Commit 5cb101fe8 introduced a format mismatch warning.  This fixes it.

Although the speed is technically unsigned, the range is such that
signedness is a don't care, and treating it as signed makes for a less
verbose cast.

TESTED:
Now builds without warning.

+++ drivers/drivers.c
@@ -1082,7 +1082,7 @@ static bool mtk3301_speed_switcher(struct gps_device_t* session,
     int i;
     char buf[BUFSIZ] = {0};
 
-    (void)snprintf(buf, sizeof(buf), "$PQBAUD,W,%u", speed);
+    (void)snprintf(buf, sizeof(buf), "$PQBAUD,W,%d", (int) speed);
 
     // Calling PQBAUD fails to set the receiver's baud rate
     // some of the time.  Sending it twice seems to work every time.
